refactor: moved header components to header folder
