+++ sites/geohub/src/routes/(map)/dashboards/electricity/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Header from '$components/Header.svelte';
+	import Header from '$components/header/Header.svelte';
 	import { AdminControlOptions, MapStyles } from '$lib/config/AppConfig';
 	import MaplibreCgazAdminControl from '@undp-data/cgaz-admin-tool';
 	import '@undp-data/cgaz-admin-tool/dist/maplibre-cgaz-admin-control.css';
