+++ sites/geohub/src/routes/(map)/map/+layout.svelte
@@ -3,7 +3,7 @@
 	import { beforeNavigate } from '$app/navigation';
 	import { page } from '$app/stores';
 	import Content from '$components/Content.svelte';
-	import Header from '$components/Header.svelte';
+	import Header from '$components/header/Header.svelte';
 	import Notification from '$components/util/Notification.svelte';
 	import { fromLocalStorage, isStyleChanged, storageKeys, toLocalStorage } from '$lib/helper';
 	import type { DashboardMapStyle, Layer, SidebarPosition } from '$lib/types';
